sequences with scheduled trigger

+++ test/model/model.go
@@ -1,14 +1,14 @@
 package model
 
 type InitMeasured struct {
-	MoistLimit      float32 `json:"moistLimit"`
-	WaterLevelLimit float32 `json:"waterLevelLimit"`
+	MoistLimit      float64 `json:"moistLimit"`
+	WaterLevelLimit float64 `json:"waterLevelLimit"`
 }
 
 type LiveMeasure struct {
-	Moist float32 `json:"moist"`
-	Hum   float32 `json:"hum"`
-	Temp  float32 `json:"temp"`
+	Moist float64 `json:"moist"`
+	Hum   float64 `json:"hum"`
+	Temp  float64 `json:"temp"`
 }
 
 type LiveNotify struct {
